+++ src/main/java/insane96mcp/iguanatweaksexpanded/data/generator/ITERecipeProvider.java
@@ -506,9 +506,12 @@ public class ITERecipeProvider extends RecipeProvider implements IConditionBuild
                 .requires(Ingredient.of(Items.GLASS_BOTTLE), 3)
                 .unlockedBy("has_milk", has(Items.MILK_BUCKET))
                 .save(writer);
-        SimpleCookingRecipeBuilder.smelting(Ingredient.of(FoodDrinks.MILK_BOTTLE.get()), RecipeCategory.FOOD, FoodDrinks.CHEESE_SLICE.get(), 0.35f, 600)
+        SimpleCookingRecipeBuilder.smelting(Ingredient.of(FoodDrinks.MILK_BOTTLE.get()), RecipeCategory.FOOD, FoodDrinks.CHEESE_SLICE.get(), 0.35f, 200)
                 .unlockedBy("has_milk", has(FoodDrinks.MILK_BOTTLE.get()))
-                .save(writer);
+                .save(writer, IguanaTweaksExpanded.RESOURCE_PREFIX + "cheese_slice_from_smelting");
+        SimpleCookingRecipeBuilder.smoking(Ingredient.of(FoodDrinks.MILK_BOTTLE.get()), RecipeCategory.FOOD, FoodDrinks.CHEESE_SLICE.get(), 0.35f, 100)
+                .unlockedBy("has_milk", has(FoodDrinks.MILK_BOTTLE.get()))
+                .save(writer, IguanaTweaksExpanded.RESOURCE_PREFIX + "cheese_slice_from_smoking");
 
         ShapelessRecipeBuilder.shapeless(RecipeCategory.TOOLS, CoalCharcoal.FIRESTARTER.get())
                 .requires(Items.FLINT, 2)