+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/hungerhealth/fooddrinks/MilkBottle.java
@@ -1,0 +1,77 @@
+package insane96mcp.iguanatweaksexpanded.module.hungerhealth.fooddrinks;
+
+import net.minecraft.advancements.CriteriaTriggers;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.stats.Stats;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResultHolder;
+import net.minecraft.world.effect.MobEffectInstance;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.*;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.gameevent.GameEvent;
+
+import java.util.List;
+
+public class MilkBottle extends Item {
+    private static final int DRINK_DURATION = 20;
+
+    public MilkBottle(Properties pProperties) {
+        super(pProperties);
+    }
+
+    @Override
+    public ItemStack finishUsingItem(ItemStack stack, Level level, LivingEntity livingEntity) {
+        Player player = livingEntity instanceof Player ? (Player)livingEntity : null;
+        if (player instanceof ServerPlayer) {
+            CriteriaTriggers.CONSUME_ITEM.trigger((ServerPlayer)player, stack);
+        }
+
+        if (!level.isClientSide) {
+            List<MobEffectInstance> mobEffectInstances = livingEntity.getActiveEffects()
+                    .stream()
+                    //TODO Check for Milk bucket and not if empty
+                    .filter(mobEffectInstance -> !mobEffectInstance.getEffect().getCurativeItems().isEmpty()).toList();
+            if (!mobEffectInstances.isEmpty()) {
+                int randomEffect = level.random.nextInt(mobEffectInstances.size());
+                livingEntity.removeEffect(mobEffectInstances.get(randomEffect).getEffect());
+            }
+        }
+
+        if (player != null) {
+            player.awardStat(Stats.ITEM_USED.get(this));
+            if (!player.getAbilities().instabuild) {
+                stack.shrink(1);
+            }
+        }
+
+        if (player == null || !player.getAbilities().instabuild) {
+            if (stack.isEmpty()) {
+                return new ItemStack(Items.GLASS_BOTTLE);
+            }
+
+            if (player != null) {
+                player.getInventory().add(new ItemStack(Items.GLASS_BOTTLE));
+            }
+        }
+
+        livingEntity.gameEvent(GameEvent.DRINK);
+        return stack;
+    }
+
+    @Override
+    public InteractionResultHolder<ItemStack> use(Level pLevel, Player pPlayer, InteractionHand pUsedHand) {
+        return ItemUtils.startUsingInstantly(pLevel, pPlayer, pUsedHand);
+    }
+
+    @Override
+    public int getUseDuration(ItemStack pStack) {
+        return DRINK_DURATION;
+    }
+
+    @Override
+    public UseAnim getUseAnimation(ItemStack pStack) {
+        return UseAnim.DRINK;
+    }
+}